Added filtering by symbol to inactive tokens
